{"schema":"decodex/commit/1","summary":"Modularize frozen transition runtime","authority":"manual"}
