+++ packages/rsnap-overlay/src/overlay/tests/self_capture_runtime.rs
@@ -658,8 +658,9 @@ fn window_matte_capture_without_live_preview_escalates_to_hidden_fallback_after_
 	assert!(!session.capture_windows_hidden);
 	assert!(session.state.frozen_display_image.is_none());
 
-	session.frozen_transition_started_at =
-		Some(Instant::now() - DISPLAY_FIRST_FREEZE_LIVE_TIMEOUT - Duration::from_millis(1));
+	session.debug_set_frozen_transition_started_at(Some(
+		Instant::now() - DISPLAY_FIRST_FREEZE_LIVE_TIMEOUT - Duration::from_millis(1),
+	));
 
 	let _ = session.about_to_wait();
 
@@ -761,8 +762,9 @@ fn background_capture_without_live_snapshot_escalates_to_hidden_fallback_after_t
 	assert!(!session.capture_windows_hidden);
 	assert!(session.state.frozen_display_image.is_none());
 
-	session.frozen_transition_started_at =
-		Some(Instant::now() - DISPLAY_FIRST_FREEZE_LIVE_TIMEOUT - Duration::from_millis(1));
+	session.debug_set_frozen_transition_started_at(Some(
+		Instant::now() - DISPLAY_FIRST_FREEZE_LIVE_TIMEOUT - Duration::from_millis(1),
+	));
 
 	let _ = session.about_to_wait();
 
@@ -786,9 +788,9 @@ fn hidden_fallback_capture_response_commits_frozen_images_while_mode_is_still_li
 		None,
 		Some(cursor),
 	);
-
-	session.frozen_transition_started_at =
-		Some(Instant::now() - DISPLAY_FIRST_FREEZE_LIVE_TIMEOUT - Duration::from_millis(1));
+	session.debug_set_frozen_transition_started_at(Some(
+		Instant::now() - DISPLAY_FIRST_FREEZE_LIVE_TIMEOUT - Duration::from_millis(1),
+	));
 
 	let _ = session.about_to_wait();
